Update for out of the OpenCPN source tree build
I made some updates for building the plugin out of the source tree of
OpenCPN with reference to the WMM plugin from Pavel.
Also some minor updates to get it running. I tested it on OSX and
Linux, but I don't have a Windows environment, so I didn't test it on
Windows.
For OSX the Layout files are not copied in the package, so you have to
manually get those.
I don't know if it would be possible to include those files in the
package, but I don't know enough about the Packages program how to do
that.