Added VWT and VWR NMEA messages
Added VWT and VWR NMEA messages.
Changed the layout a bit and renamed Winddirection and Windspeed to TWA
and TWS and also added new columns AWA and AWS.
One thing I could not fix in the code is the layout of Textblocks.xml
file in the data directory of Logbook.
So when you apply this fix you should delete the Textblocks.xml file
and recreate the entries.