Updates
Changed the Sign column to Status column (Sign was not really clear I
think)

Changed the help system by moving the files from the user area to the
data area of the system, so now we can do updates to the help files
without reloading the Layout files.
Now the Layout zip file only contains the HTML and ODT layout files.

Updated the Helpfile to reflect the changes we made to Logbookkonni.
Still not completely done here.